debug btcstk mod
