+++ x/btcstaking/keeper/voting_power_table_test.go
@@ -2,7 +2,6 @@ package keeper_test
 
 import (
 	"math/rand"
-	"sort"
 	"testing"
 
 	"github.com/golang/mock/gomock"
@@ -10,6 +9,7 @@ import (
 
 	"github.com/babylonlabs-io/babylon/testutil/datagen"
 	btclctypes "github.com/babylonlabs-io/babylon/x/btclightclient/types"
+	"github.com/babylonlabs-io/babylon/x/btcstaking/testutil"
 	"github.com/babylonlabs-io/babylon/x/btcstaking/types"
 )
 
@@ -24,9 +24,7 @@ func FuzzVotingPowerTable(f *testing.F) {
 		// mock BTC light client and BTC checkpoint modules
 		btclcKeeper := types.NewMockBTCLightClientKeeper(ctrl)
 		btccKeeper := types.NewMockBtcCheckpointKeeper(ctrl)
-		finalityKeeper := types.NewMockFinalityKeeper(ctrl)
-		finalityKeeper.EXPECT().HasTimestampedPubRand(gomock.Any(), gomock.Any(), gomock.Any()).Return(true).AnyTimes()
-		h := NewHelper(t, btclcKeeper, btccKeeper, finalityKeeper)
+		h := testutil.NewHelper(t, btclcKeeper, btccKeeper)
 
 		// set all parameters
 		covenantSKs, _ := h.GenAndApplyParams(r)
@@ -158,288 +156,5 @@ func FuzzVotingPowerTable(f *testing.F) {
 		activatedHeight2, err := h.BTCStakingKeeper.GetBTCStakingActivatedHeight(h.Ctx)
 		require.NoError(t, err)
 		require.Equal(t, activatedHeight, activatedHeight2)
-	})
-}
-
-func FuzzVotingPowerTable_ActiveFinalityProviders(f *testing.F) {
-	datagen.AddRandomSeedsToFuzzer(f, 10)
-
-	f.Fuzz(func(t *testing.T, seed int64) {
-		r := rand.New(rand.NewSource(seed))
-		ctrl := gomock.NewController(t)
-		defer ctrl.Finish()
-
-		// mock BTC light client and BTC checkpoint modules
-		btclcKeeper := types.NewMockBTCLightClientKeeper(ctrl)
-		btccKeeper := types.NewMockBtcCheckpointKeeper(ctrl)
-		finalityKeeper := types.NewMockFinalityKeeper(ctrl)
-		h := NewHelper(t, btclcKeeper, btccKeeper, finalityKeeper)
-
-		// set all parameters
-		covenantSKs, _ := h.GenAndApplyParams(r)
-		changeAddress, err := datagen.GenRandomBTCAddress(r, h.Net)
-		h.NoError(err)
-
-		// generate a random batch of finality providers, each with a BTC delegation with random power
-		fpsWithMeta := []*types.FinalityProviderDistInfo{}
-		numFps := datagen.RandomInt(r, 300) + 1
-		noTimestampedFps := map[string]bool{}
-		for i := uint64(0); i < numFps; i++ {
-			// generate finality provider
-			_, _, fp := h.CreateFinalityProvider(r)
-
-			// delegate to this finality provider
-			stakingValue := datagen.RandomInt(r, 100000) + 100000
-			delSK, _, err := datagen.GenRandomBTCKeyPair(r)
-			h.NoError(err)
-			stakingTxHash, delMsg, del, btcHeaderInfo, inclusionProof, err := h.CreateDelegation(
-				r,
-				delSK,
-				fp.BtcPk.MustToBTCPK(),
-				changeAddress.EncodeAddress(),
-				int64(stakingValue),
-				1000,
-				0,
-				0,
-				true,
-			)
-			h.NoError(err)
-			h.CreateCovenantSigs(r, covenantSKs, delMsg, del)
-			h.AddInclusionProof(stakingTxHash, btcHeaderInfo, inclusionProof)
-
-			// 30 percent not have timestamped randomness, which causes
-			// zero voting power in the table
-			fpDistInfo := &types.FinalityProviderDistInfo{BtcPk: fp.BtcPk, TotalBondedSat: stakingValue}
-			if r.Intn(10) <= 2 {
-				finalityKeeper.EXPECT().HasTimestampedPubRand(gomock.Any(), fp.BtcPk, gomock.Any()).Return(false).AnyTimes()
-				noTimestampedFps[fp.BtcPk.MarshalHex()] = true
-				fpDistInfo.IsTimestamped = false
-			} else {
-				finalityKeeper.EXPECT().HasTimestampedPubRand(gomock.Any(), fp.BtcPk, gomock.Any()).Return(true).AnyTimes()
-				fpDistInfo.IsTimestamped = true
-			}
-
-			// record voting power
-			fpsWithMeta = append(fpsWithMeta, fpDistInfo)
-		}
-
-		maxActiveFpsParam := h.BTCStakingKeeper.GetParams(h.Ctx).MaxActiveFinalityProviders
-		// get a map of expected active finality providers
-		types.SortFinalityProvidersWithZeroedVotingPower(fpsWithMeta)
-		expectedActiveFps := fpsWithMeta[:min(uint32(len(fpsWithMeta)-len(noTimestampedFps)), maxActiveFpsParam)]
-		expectedActiveFpsMap := map[string]uint64{}
-		for _, fp := range expectedActiveFps {
-			expectedActiveFpsMap[fp.BtcPk.MarshalHex()] = fp.TotalBondedSat
-		}
-
-		// record voting power table
-		babylonHeight := datagen.RandomInt(r, 10) + 1
-		h.SetCtxHeight(babylonHeight)
-		h.BTCLightClientKeeper.EXPECT().GetTipInfo(gomock.Eq(h.Ctx)).Return(&btclctypes.BTCHeaderInfo{Height: 30}).AnyTimes()
-		err = h.BTCStakingKeeper.BeginBlocker(h.Ctx)
-		require.NoError(t, err)
-
-		// only finality providers in expectedActiveFpsMap have voting power
-		for _, fp := range fpsWithMeta {
-			power := h.BTCStakingKeeper.GetVotingPower(h.Ctx, fp.BtcPk.MustMarshal(), babylonHeight)
-			if expectedPower, ok := expectedActiveFpsMap[fp.BtcPk.MarshalHex()]; ok {
-				require.Equal(t, expectedPower, power)
-			} else {
-				require.Zero(t, power)
-			}
-		}
-
-		// also, get voting power table and assert there is
-		// min(len(expectedActiveFps), MaxActiveFinalityProviders) active finality providers
-		powerTable := h.BTCStakingKeeper.GetVotingPowerTable(h.Ctx, babylonHeight)
-		expectedNumActiveFps := len(expectedActiveFpsMap)
-		if expectedNumActiveFps > int(maxActiveFpsParam) {
-			expectedNumActiveFps = int(maxActiveFpsParam)
-		}
-		require.Len(t, powerTable, expectedNumActiveFps)
-		// assert consistency of voting power
-		for pkHex, expectedPower := range expectedActiveFpsMap {
-			require.Equal(t, powerTable[pkHex], expectedPower)
-		}
-	})
-}
-
-func FuzzVotingPowerTable_ActiveFinalityProviderRotation(f *testing.F) {
-	datagen.AddRandomSeedsToFuzzer(f, 10)
-
-	f.Fuzz(func(t *testing.T, seed int64) {
-		r := rand.New(rand.NewSource(seed))
-		ctrl := gomock.NewController(t)
-		defer ctrl.Finish()
-
-		// mock BTC light client and BTC checkpoint modules
-		btclcKeeper := types.NewMockBTCLightClientKeeper(ctrl)
-		btccKeeper := types.NewMockBtcCheckpointKeeper(ctrl)
-		finalityKeeper := types.NewMockFinalityKeeper(ctrl)
-		finalityKeeper.EXPECT().HasTimestampedPubRand(gomock.Any(), gomock.Any(), gomock.Any()).Return(true).AnyTimes()
-		h := NewHelper(t, btclcKeeper, btccKeeper, finalityKeeper)
-
-		// set all parameters
-		covenantSKs, _ := h.GenAndApplyParams(r)
-		// set random number of max number of finality providers
-		// in order to cover cases that number of finality providers is more or
-		// less than `MaxActiveFinalityProviders`
-		bsParams := h.BTCStakingKeeper.GetParams(h.Ctx)
-		bsParams.MaxActiveFinalityProviders = uint32(datagen.RandomInt(r, 20) + 10)
-		err := h.BTCStakingKeeper.SetParams(h.Ctx, bsParams)
-		h.NoError(err)
-		// change address
-		changeAddress, err := datagen.GenRandomBTCAddress(r, h.Net)
-		h.NoError(err)
-
-		numFps := datagen.RandomInt(r, 20) + 10
-		numActiveFPs := int(min(numFps, uint64(bsParams.MaxActiveFinalityProviders)))
-
-		/*
-			Generate a random batch of finality providers, each with a BTC delegation
-			with random voting power.
-			Then, assert voting power table
-		*/
-		fpsWithMeta := []*types.FinalityProviderWithMeta{}
-		for i := uint64(0); i < numFps; i++ {
-			// generate finality provider
-			// generate and insert new finality provider
-			_, fpPK, fp := h.CreateFinalityProvider(r)
-
-			// create BTC delegation and add covenant signatures to activate it
-			stakingValue := datagen.RandomInt(r, 100000) + 100000
-			delSK, _, err := datagen.GenRandomBTCKeyPair(r)
-			h.NoError(err)
-			stakingTxHash, delMsg, del, btcHeaderInfo, inclusionProof, err := h.CreateDelegation(
-				r,
-				delSK,
-				fpPK,
-				changeAddress.EncodeAddress(),
-				int64(stakingValue),
-				1000,
-				0,
-				0,
-				true,
-			)
-			h.NoError(err)
-			h.CreateCovenantSigs(r, covenantSKs, delMsg, del)
-			h.AddInclusionProof(stakingTxHash, btcHeaderInfo, inclusionProof)
-
-			// record voting power
-			fpsWithMeta = append(fpsWithMeta, &types.FinalityProviderWithMeta{
-				BtcPk:       fp.BtcPk,
-				VotingPower: stakingValue,
-			})
-		}
-
-		// record voting power table
-		babylonHeight := datagen.RandomInt(r, 10) + 1
-		h.Ctx = datagen.WithCtxHeight(h.Ctx, babylonHeight)
-		h.BTCLightClientKeeper.EXPECT().GetTipInfo(gomock.Eq(h.Ctx)).Return(&btclctypes.BTCHeaderInfo{Height: 30}).AnyTimes()
-		err = h.BTCStakingKeeper.BeginBlocker(h.Ctx)
-		h.NoError(err)
-
-		// assert that only top `min(MaxActiveFinalityProviders, numFPs)` finality providers have voting power
-		sort.SliceStable(fpsWithMeta, func(i, j int) bool {
-			return fpsWithMeta[i].VotingPower > fpsWithMeta[j].VotingPower
-		})
-		for i := 0; i < numActiveFPs; i++ {
-			votingPower := h.BTCStakingKeeper.GetVotingPower(h.Ctx, *fpsWithMeta[i].BtcPk, babylonHeight)
-			require.Equal(t, fpsWithMeta[i].VotingPower, votingPower)
-		}
-		for i := numActiveFPs; i < int(numFps); i++ {
-			votingPower := h.BTCStakingKeeper.GetVotingPower(h.Ctx, *fpsWithMeta[i].BtcPk, babylonHeight)
-			require.Zero(t, votingPower)
-		}
-
-		/*
-			Delegate more tokens to some existing finality providers
-			, and create some new finality providers
-			Then assert voting power table again
-		*/
-		// delegate more tokens to some existing finality providers
-		for i := uint64(0); i < numFps; i++ {
-			if !datagen.OneInN(r, 2) {
-				continue
-			}
-
-			stakingValue := datagen.RandomInt(r, 100000) + 100000
-			fpBTCPK := fpsWithMeta[i].BtcPk
-			delSK, _, err := datagen.GenRandomBTCKeyPair(r)
-			h.NoError(err)
-			stakingTxHash, delMsg, del, btcHeaderInfo, inclusionProof, err := h.CreateDelegation(
-				r,
-				delSK,
-				fpBTCPK.MustToBTCPK(),
-				changeAddress.EncodeAddress(),
-				int64(stakingValue),
-				1000,
-				0,
-				0,
-				true,
-			)
-			h.NoError(err)
-			h.CreateCovenantSigs(r, covenantSKs, delMsg, del)
-			h.AddInclusionProof(stakingTxHash, btcHeaderInfo, inclusionProof)
-
-			// accumulate voting power for this finality provider
-			fpsWithMeta[i].VotingPower += stakingValue
-
-			break
-		}
-		// create more finality providers
-		numNewFps := datagen.RandomInt(r, 20) + 10
-		numFps += numNewFps
-		numActiveFPs = int(min(numFps, uint64(bsParams.MaxActiveFinalityProviders)))
-		for i := uint64(0); i < numNewFps; i++ {
-			// generate finality provider
-			// generate and insert new finality provider
-			_, fpPK, fp := h.CreateFinalityProvider(r)
-
-			// create BTC delegation and add covenant signatures to activate it
-			stakingValue := datagen.RandomInt(r, 100000) + 100000
-			delSK, _, err := datagen.GenRandomBTCKeyPair(r)
-			h.NoError(err)
-			stakingTxHash, delMsg, del, btcHeaderInfo, inclusionProof, err := h.CreateDelegation(
-				r,
-				delSK,
-				fpPK,
-				changeAddress.EncodeAddress(),
-				int64(stakingValue),
-				1000,
-				0,
-				0,
-				true,
-			)
-			h.NoError(err)
-			h.CreateCovenantSigs(r, covenantSKs, delMsg, del)
-			h.AddInclusionProof(stakingTxHash, btcHeaderInfo, inclusionProof)
-
-			// record voting power
-			fpsWithMeta = append(fpsWithMeta, &types.FinalityProviderWithMeta{
-				BtcPk:       fp.BtcPk,
-				VotingPower: stakingValue,
-			})
-		}
-
-		// record voting power table
-		babylonHeight += 1
-		h.Ctx = datagen.WithCtxHeight(h.Ctx, babylonHeight)
-		h.BTCLightClientKeeper.EXPECT().GetTipInfo(gomock.Eq(h.Ctx)).Return(&btclctypes.BTCHeaderInfo{Height: 30}).AnyTimes()
-		err = h.BTCStakingKeeper.BeginBlocker(h.Ctx)
-		h.NoError(err)
-
-		// again, assert that only top `min(MaxActiveFinalityProviders, numFPs)` finality providers have voting power
-		sort.SliceStable(fpsWithMeta, func(i, j int) bool {
-			return fpsWithMeta[i].VotingPower > fpsWithMeta[j].VotingPower
-		})
-		for i := 0; i < numActiveFPs; i++ {
-			votingPower := h.BTCStakingKeeper.GetVotingPower(h.Ctx, *fpsWithMeta[i].BtcPk, babylonHeight)
-			require.Equal(t, fpsWithMeta[i].VotingPower, votingPower)
-		}
-		for i := numActiveFPs; i < int(numFps); i++ {
-			votingPower := h.BTCStakingKeeper.GetVotingPower(h.Ctx, *fpsWithMeta[i].BtcPk, babylonHeight)
-			require.Zero(t, votingPower)
-		}
 	})
 }
